🚚 move: relocate product-category-list related files to /bridge/productcategorylist package
