+++ src/main/java/com/jokim/sivillage/api/bridge/productcategorylist/infrastructure/ProductCategoryListRepository.java
@@ -1,6 +1,6 @@
-package com.jokim.sivillage.api.bridge.infrastructure;
+package com.jokim.sivillage.api.bridge.productcategorylist.infrastructure;
 
-import com.jokim.sivillage.api.bridge.domain.ProductCategoryList;
+import com.jokim.sivillage.api.bridge.productcategorylist.domain.ProductCategoryList;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductCategoryListRepository extends JpaRepository<ProductCategoryList, Long> {
